+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Look in your home directory for goldapi.io.txt.
 - License files.
 - The demo ledger to the debian package.
 - A copyright date.